Added authorization token and full reconnect with signed players

+++ AvoidContactServer/Networking/Enums/Results/SignInResult.cs
@@ -1,6 +1,6 @@
 ﻿namespace AvoidContactServer.Networking.Enums.Results
 {
-    public enum SignInResult : ushort
+    public enum SignInResult : byte
     {
         Success = 0,
         WrongLoginOrPassword = 1,
